src/commands/AddCommand.js -> commands/add/index.js
