+++ commands/add/__tests__/__fixtures__/basic/packages/package-4/package.json
