+++ commands/add/__tests__/__fixtures__/existing/lerna.json
