+++ commands/add/__tests__/__fixtures__/existing/package.json
