+++ commands/add/__tests__/__fixtures__/existing/packages/package-1/package.json
