+++ commands/add/__tests__/__fixtures__/existing/packages/package-2/package.json
