+++ commands/add/__tests__/__fixtures__/existing/packages/package-3/package.json
