+++ commands/add/__tests__/__fixtures__/unchanged/packages/package-2/package.json
