+++ commands/add/lib/not-satisfied-message.js
@@ -1,0 +1,12 @@
+"use strict";
+
+const dedent = require("dedent");
+
+module.exports = notSatisfiedMessage;
+
+function notSatisfiedMessage(unsatisfied) {
+  return dedent`
+    Requested range not satisfiable:
+    ${unsatisfied.map(u => `${u.name}@${u.versionRange} (available: ${u.version})`).join(", ")}
+  `;
+}
