+++ commands/add/package.json
@@ -14,7 +14,8 @@
   },
   "files": [
     "command.js",
-    "index.js"
+    "index.js",
+    "lib"
   ],
   "main": "index.js",
   "engines": {
@@ -30,5 +31,16 @@
   },
   "scripts": {
     "test": "echo \"Run tests from root\" && exit 1"
+  },
+  "dependencies": {
+    "@lerna/bootstrap": "file:../bootstrap",
+    "@lerna/command": "file:../../core/command",
+    "@lerna/validation-error": "file:../../core/validation-error",
+    "dedent": "^0.7.0",
+    "npm-package-arg": "^6.0.0",
+    "package-json": "^4.0.1",
+    "read-pkg": "^3.0.0",
+    "semver": "^5.5.0",
+    "write-pkg": "^3.1.0"
   }
 }
